+++ mountpoint-s3/tests/fs.rs
@@ -577,8 +577,13 @@ async fn test_sequential_write(write_size: usize) {
     let file_ino = dentry.attr.ino;
 
     // First let's check that we can't write it again
-    let result = fs
+    let fh = fs
         .open(file_ino, libc::S_IFREG as i32 | libc::O_WRONLY, 0)
+        .await
+        .unwrap()
+        .fh;
+    let result = fs
+        .write(file_ino, fh, offset, &[0xaa; 27], 0, 0, None)
         .await
         .expect_err("file should not be overwritable")
         .to_errno();
@@ -669,14 +674,22 @@ async fn test_duplicate_write_fails() {
     assert_eq!(dentry.attr.size, 0);
     let file_ino = dentry.attr.ino;
 
-    let _opened = fs
+    let opened = fs
         .open(file_ino, libc::S_IFREG as i32 | libc::O_WRONLY, 0)
         .await
         .unwrap();
+    _ = fs
+        .write(file_ino, opened.fh, 0, &[0xaa; 27], 0, 0, None)
+        .await
+        .expect("first write should succeed");
 
-    // Should not be allowed to open the file a second time
-    let err = fs
+    let opened = fs
         .open(file_ino, libc::S_IFREG as i32 | libc::O_WRONLY, 0)
+        .await
+        .expect("open should succeed");
+    // Should not be allowed to write the file a second time
+    let err = fs
+        .write(file_ino, opened.fh, 0, &[0xaa; 27], 0, 0, None)
         .await
         .expect_err("should not be able to write twice")
         .to_errno();